Upload flag defaults to false
With the new meaning of the upload flag, we should not do it by default,
as that would really defeat the purpose of listening to the file.
Instead, the flag will often be used by the client as a seperate
execution and upload the file while the parser is still running.

Fix #1