Parse Rank Updates
The game, annoyingly, only asks for the player rank at the beginning and
then rarely again. This means that when you play for a bit and the rank
changes, we won't parse it again until you restart the client. Now, we
run through the entire log looking for rank updates and update the
original rank we find, sending only the final version to the server.